+++ src/Objects/CalendarMelanie.php
@@ -448,13 +448,16 @@ class CalendarMelanie extends MagicObject implements IObjectMelanie {
 	/**
 	 * Recupère le Tag associé à l'agenda
 	 * need: $this->calendar_id
+	 * 
+	 * @param boolean $cache Utilisation du ctag en cache ?
+	 * 
 	 * @return string
 	 */
-	public function getCTag() {
+	public function getCTag($cache = true) {
 		M2Log::Log(M2Log::LEVEL_DEBUG, $this->get_class."->getCTag()");
 		if (!isset($this->id)) return false;
 
-		if (!isset($this->ctag)) {
+		if (!isset($this->ctag) || !$cache) {
 			// Params
 			$params = [MappingMce::$Data_Mapping[$this->objectType]['id'][MappingMce::name] => $this->id];
 
@@ -464,6 +467,30 @@ class CalendarMelanie extends MagicObject implements IObjectMelanie {
 			M2Log::Log(M2Log::LEVEL_DEBUG, $this->get_class."->getCTag() this->ctag: " . $this->ctag);
 		}
 		return $this->ctag;
+	}
+
+	/**
+	 * Recupère le SyncToken associé à l'agenda
+	 * need: $this->calendar_id
+	 * 
+	 * @param boolean $cache Utilisation du synctoken en cache ?
+	 * 
+	 * @return string
+	 */
+	public function getSyncToken($cache = true) {
+		M2Log::Log(M2Log::LEVEL_DEBUG, $this->get_class."->getSyncToken()");
+		if (!isset($this->id)) return false;
+
+		if (!isset($this->synctoken) || !$cache) {
+			// Params
+			$params = [MappingMce::$Data_Mapping[$this->objectType]['id'][MappingMce::name] => $this->id];
+
+			// Récupération du tag
+			Sql\Sql::GetInstance()->executeQueryToObject(Sql\SqlCalendarRequests::getSyncToken, $params, $this);
+			if (!isset($this->synctoken)) $this->synctoken = 0;
+			M2Log::Log(M2Log::LEVEL_DEBUG, $this->get_class."->getSyncToken() this->synctoken: " . $this->synctoken);
+		}
+		return $this->synctoken;
 	}
 
 	/**